requirements.txt
flask>=3.0.0
anthropic>=0.25.0
openai>=1.30.0
librosa>=0.10.0
numpy>=1.26.0
pydub>=0.25.1
python-dotenv>=1.0.0
werkzeug>=3.0.0
soundfile>=0.12.1
